### 2025-02-21 #7835 @editorialbot
[PRE REVIEW]: hermes-rheo: An open-source Python package for rheological data analysis
**Submitting author:** @ale-perego (<a href="http://orcid.org/0000-0002-0570-3210">Alessandro Perego</a>)
**Repository:** https://github.com/3mcloud/hermes-rheo/
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.1
