+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = hermes_rheo
-version = 1.2.1
+version = 1.2.2
 description = piblin transforms tailored for advanced rheological analysis
 long_description = file: README.md
 long_description_content_type = text/markdown
