+++ src/hermes_rheo/transforms/rheo_analysis.py
@@ -1,9 +1,8 @@
 from typing import Tuple, Dict, List, Optional, Any
-import numpy as np
 import math
 import copy
 from scipy.fftpack import fft, fftshift
-
+import numpy as np
 from piblin.transform.abc.measurement_transform import MeasurementTransform
 import piblin.data.datasets.abc.split_datasets.one_dimensional_composite_dataset as one_d_composite_dataset
 
@@ -28,10 +27,12 @@ class RheoAnalysis(MeasurementTransform):
 
     """
 
-    def __init__(self, owchirp_waiting_time='before_signal', cutoff_points=0, *args, **kwargs):
+    def __init__(self, owchirp_waiting_time='before_signal', cutoff_points=0, filter_criterion="endpoints", *args,
+                 **kwargs):
         super().__init__(*args, **kwargs)
         self.owchirp_waiting_time = owchirp_waiting_time
         self.cutoff_points = cutoff_points
+        self.filter_criterion = filter_criterion  # Store filter_criterion as an instance variable
 
     @staticmethod
     def prepare_owchirp_data(dataset: one_d_composite_dataset, time: str, strain: str, stress: str, temperature: str,
@@ -256,23 +257,26 @@ class RheoAnalysis(MeasurementTransform):
 
         return corrected_signal
 
-    from typing import List, Tuple, Optional
-    import numpy as np
-
-    def select_best_filter_method(self, t_w: float, time: List[float], signal: List[float]) -> Tuple[np.ndarray, str]:
+    def select_best_filter_method(self, t_w: float, time: List[float], signal: List[float],
+                                  criterion: str = "symmetry") -> Tuple[np.ndarray, str]:
         """
         Iterates over different filtering methods, applies each to the signal, and evaluates the corrected signal.
-        The best filtering method is selected based on the lowest sum of absolute values of the corrected signal,
-        indicating the most symmetrical signal around zero.
+        The best filtering method is selected based on the specified criterion.
 
         Args:
             t_w (float): The time threshold used for filtering.
             time (List[float]): A list of time values corresponding to the signal.
             signal (List[float]): A list of signal values to be filtered.
+            criterion (str): The criterion for selecting the best method. Options:
+                - "symmetry": Selects the method with the most symmetrical signal (lowest sum of absolute values).
+                - "DMA": Selects the method with minimal bias for DMA instruments (combined start/end avg + std dev).
 
         Returns:
             Tuple[np.ndarray, str]: A tuple containing the best corrected signal as a NumPy array
                                     and the name of the best method as a string.
+
+        Raises:
+            ValueError: If an invalid criterion is provided.
         """
         # Ensure time and signal are converted to np.ndarray for consistent operations
         time = np.asarray(time)
@@ -280,66 +284,33 @@ class RheoAnalysis(MeasurementTransform):
 
         methods = ['tw', 'chirp', 'all', 'none']
         best_method = None
-        min_sum_abs = float('inf')
+        min_score = float('inf')
 
-        for method in methods:
-            corrected_signal = self.filter_signal(t_w, time, signal, method)
-            if corrected_signal is not None:
-                sum_abs = np.sum(np.abs(corrected_signal))
-
-                if sum_abs < min_sum_abs:
-                    min_sum_abs = sum_abs
-                    best_method = method
-
-        best_corrected_signal = self.filter_signal(t_w, time, signal, best_method)
-        return best_corrected_signal, best_method
-
-    def select_best_filter_method_DMA(self, t_w: float, time: List[float], signal: List[float]) -> Tuple[
-        np.ndarray, str]:
-        """
-        Iterates over different filtering methods and selects the best one for DMA instruments.
-        The method with the best metric, calculated by combining the average of the start and end signal values and the
[…]
