### 2025-03-21 70bf05d
F polish rheo analysis (#29)
* v1.2.3
