### 2025-03-21 765683c
added idea
