### 2025-03-21 8f53b41
Develop (#32)
* Fixed Frequency dictionary and add readthedocs.yml

* Upgraded version to start deploying docs

* Updating docs

* More docs

* Updated docs

* Modified azure pipeline to account for modern python versions

* Updaded azure pipeline to reflect OS versions

* Updated python requirements up to 3.12

* 'Test'

* 'PolishDocs'

* 'Updated_README_file'

* v1.0.12 (#13)

* 'Updaded_setup.cfg'

* Updaded conftest

---------

Co-authored-by: Alessandro Perego <aperego@mmm.com>

* F options docs (#15)

* fix: made sphinx doc requirement

* fix: position of options requirement in setup fil
[…]
