### 2025-03-21 7495f11
Develop (#35)
* Stop ignoring source files

* Fix git (#34)

* v1.2.6

---------

Co-authored-by: Alessandro Perego <aperego@mmm.com>
