### 2025-03-21 621abff
Develop (#40)
* add comment to RheoAnalysis (#39)

* v1.2.7

---------

Co-authored-by: Alessandro Perego <aperego@mmm.com>
