### 2025-06-24 #7835 @kyleniemeyer
Hi @diehlpk, could you edit this submission? Your expertise in solid mechanics makes you one of the closer fits topic-wise. Thanks!
