### 2025-06-26 #7835 @xcheng86
Thanks for the invitation. I am willing to take the time to review this paper.
