### 2025-07-08 #7835 @adambeallwsp
> Hi [@adambeall](https://github.com/adambeall) do you have time to review this paper?

Hi @diehlpk, apologies I am currently overcommitted and won't be able to review this paper
