### 2025-08-24 #8753 @ramisetti
Thanks for reaching out. Unfortunately, I’m unable to review at this time due to other commitments.

> On 22-Aug-2025, at 3:38 AM, Patrick Diehl ***@***.***> wrote:
>
>
> diehlpk
>  left a comment
> (openjournals/joss-reviews#8753)
>  <https://github.com/openjournals/joss-reviews/issues/8753#issuecomment-3212213913>
> Hi @ramisetti <https://github.com/ramisetti> do you have time to review this paper?
>
