### 2025-10-31 #8753 @ale-perego
@diehlpk The tests are automated upon deployment : https://github.com/3mcloud/hermes-rheo/blob/main/azure-pipelines.yml
Not sure what other types of automation are you refering.

Concerning the reviewers, I can suggest https://github.com/krlennon
