### 2025-11-26 #8753 @dhelmrich
@diehlpk The software package is quite mature in its design and usability. The documentation is good, the clearest instructional content are the jupyter notebooks. It is more user-friendly than average for scientific software, particularly given its field.

The lack of self-reference is concerning and I have already mentioned that I believe that the JOSS submission infringes on Springer's copyright regarding this software package. At the very least, the citation needs to be included referring to the Springer publication. That being said, I sympathize with the authors regarding the fact that this review process has taken a very long time, leading inadvertently to co-publication.

I concluded the checklist and recommend publication after the editors and authors together address the licence issue. I have left the references checkmark open because of the ref issue, but since it is a larger question I am treating it differently to the normal JOSS change requests.
