+++ paper/paper.bib
@@ -87,3 +87,12 @@
     year={2023},
     pages={e13}
 }
+
+@article{peregoowch2025,
+    title={Evaluation of optimally windowed chirp signals in industrial rheological measurements: method development and data processing},
+    volume={64}, number={8}, DOI={10.1007/s00397-025-01511-0},
+    journal={Rheologica Acta},
+    author={Perego, Alessandro and Vadillo, Damien C. and Mills, Matthew J. L. and Das, Mohua and McKinley, Gareth H.},
+    year={2025},
+    pages={391--406}
+}
