### 2026-01-03 f7cad7f
CI: replace retired ubuntu-20.04 and macOS-13 images
