+++ azure-pipelines.yml
@@ -20,40 +20,32 @@ jobs:
         Python39-windows-2022:
           vmimage: 'windows-2022'
           python.version: '3.9'
-        Python39-ubuntu-20.04:
-          vmimage: 'ubuntu-20.04'
-          python.version: '3.9'
         Python39-ubuntu-22.04:
           vmimage: 'ubuntu-22.04'
           python.version: '3.9'
-        Python39-macOS-13:
-          vmimage: 'macOS-13'
+        Python39-macOS-14:
+          vmimage: 'macOS-14'
           python.version: '3.9'
+
         Python310-windows-2022:
           vmimage: 'windows-2022'
-          python.version: '3.10'
-        Python310-ubuntu-20.04:
-          vmimage: 'ubuntu-20.04'
           python.version: '3.10'
         Python310-ubuntu-22.04:
           vmimage: 'ubuntu-22.04'
           python.version: '3.10'
-        Python310-macOS-13:
-          vmimage: 'macOS-13'
+        Python310-macOS-14:
+          vmimage: 'macOS-14'
           python.version: '3.10'
+
         Python311-windows-2022:
           vmimage: 'windows-2022'
-          python.version: '3.11'
-        Python311-ubuntu-20.04:
-          vmimage: 'ubuntu-20.04'
           python.version: '3.11'
         Python311-ubuntu-22.04:
           vmimage: 'ubuntu-22.04'
           python.version: '3.11'
-        Python311-macOS-13:
-          vmimage: 'macOS-13'
+        Python311-macOS-14:
+          vmimage: 'macOS-14'
           python.version: '3.11'
-
 
     pool:
       vmImage: $(vmimage)
