### 2026-01-03 d5af18d
Bump version to 1.2.8 (#47)
