### 2026-01-03 9baa37f
Release 1.2.8 (#49)
