+++ paper/paper.bib
@@ -9,11 +9,11 @@
 
 @misc{perego2024hermesrheo,
   author = {Perego, Alessandro and Mills, Matthew J. L. and Vadillo, Damien C.},
-  title = {hermes-rheo (1.2.0)},
+  title = {hermes-rheo},
   year = {2024},
   publisher = {Zenodo},
-  doi = {10.5281/zenodo.14453081},
-  url = {https://doi.org/10.5281/zenodo.14453081}
+  doi = {10.5281/zenodo.14182224},
+  url = {https://doi.org/10.5281/zenodo.14182224}
 }
 
 @article{owchirp,
