### 2026-01-16 fe1b001
JOSS final review changes (#50)
