### 2026-01-16 acc9b3d
updaed vm Image to windows 2022
