### 2026-01-16 995c78f
Develop (#53)
* JOSS final review changes (#50)

* JOSS final review changes (#50) (#51) (#52)

* JOSS final review changes (#50) (#51)

* fixed vmImage .yml

* removed DOI from readme for now
