### 2026-01-19 #8753 @ale-perego
@diehlpk I saw there was an error with the OCID of one of the author. What is the reccomended procedure now? Should I made the modification and push the changes or do I need to tag and push a new version and archive the current one?
