### 2026-01-19 #8753 @diehlpk
Just update the paper.md in the JOSS branch. No need for a new release or DOI.
