+++ paper/paper.md
@@ -10,7 +10,7 @@ authors:
     orcid: 0000-0002-0570-3210
     affiliation: 1
   - name: Matthew J.L. Mills
-    orcid: 000-0002-7208-7542
+    orcid: 0000-0002-7208-7542
     affiliation: 1
   - name: Damien C. Vadillo
     orcid: 0000-0003-0555-745X
