### 2026-01-20 #8753 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
