### 2026-01-20 6becda7
JOSS paper edits
