### 2026-01-20 #8753 @kyleniemeyer
@ale-perego I've made some minor edits in https://github.com/3mcloud/hermes-rheo/pull/56, can you review and merge? We can then proceed to publication.
